requirements.txt
altgraph==0.17.4
build==1.0.3
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
docutils==0.20.1
grpcio==1.59.2
grpcio-testing==1.59.2
grpcio-tools==1.59.2
idna==3.4
importlib-metadata==6.8.0
jaraco.classes==3.3.0
keyring==24.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.1.0
nh3==0.2.14
packaging==23.2
pefile==2023.2.7
pkginfo==1.9.6
protobuf==4.25.0
Pygments==2.16.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
pyproject_hooks==1.0.0
python-mnist==0.7
pywin32-ctypes==0.2.2
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
setuptools==68.2.2
twine==4.0.2
urllib3==2.0.7
zipp==3.17.0
